Fix loading of the integrated too panel
Still need to sort this out more consistently but it is a permissions issue with changing the ownership of the /galaxy/server/database folder in the init containers. This ultimately leads to an incorrect (i.e., default) integrated tool panel to be loaded. Mixing init container images is a workaround because the alpine container runs as root so can set the permissions.