Remove the install_database_connection conf option
This option is needed if the installed tools (and Galaxy) are to be distributed as a unit. In our scenario where any additional tools are installed on a live instance, the install details will be saved in the main Galaxy database and there is no need for a dedicated database.

xref https://github.com/CloudVE/galaxy-kubernetes/issues/21